admin login added, jwt added, Enum added, common methods added, messages added, throw error added, parsing methods added, response added, auth middleware added

+++ src/items/controllers/item.controller.ts
@@ -1,84 +1,89 @@
 import { ItemInterface } from "../schemas/items.schema";
-import { Request, Response } from 'express';
+import { NextFunction, Response } from 'express';
 import ItemServiceModel from "../models/items.serviceModel";
-import ErrData from "../../common/http-exception";
-
+import { ErrData } from "../../common/http-exception";
+import { messages, response } from "../../middleware/util.middleware";
+import { TypeInterface as Request, toStrings } from "../../common/types.interface";
+import { Enum } from "../../common/enum";
 
 // Create item method
-export const createController = async (req: Request, res: Response) => {
+export const createItemController = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        let reqBody: ItemInterface = req.body;
-        const result = await ItemServiceModel.createItem(reqBody);
+        if (req.user.role !== Enum.adminRoleId)
+            throw ErrData(400, messages.notAuthorized(), false);
+        let Body: ItemInterface = req.body;
+        const result = await ItemServiceModel.createItem(Body);
         if (!result)
-            throw new ErrData(400, false, "Item could not be saved!");
-        // return res.status(400).send({ message: "Item could not be saved!" });
-            return res.status(200).send({ message: "Item added successfully!" });
+            throw ErrData(400, messages.itemNotSaved(), false);
+        return response(res, 200, true, "Item added successfully!", result);
     }
     catch (error) {
         console.error("Catch error:-", error);
-        res.status(500).send(error);
+        next(error);
     }
 }
 
 // Find all item list
-export const findAllItemController = async (req: Request, res: Response) => {
+export const findAllItemController = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const items = await ItemServiceModel.findAll();
-        res.status(200).send(items);
+        return response(res, 200, true, "Data found!", items);
     }
     catch (error) {
         console.error("Catch error:-", error);
-        res.status(500).send(error);
+        next(error);
     }
 }
 
 // Find item by id
-export const findItemController = async (req: Request, res: Response) => {
+export const findItemController = async (req: Request, res: Response, next: NextFunction) => {
     try {
         let id: string = (req.query.id + "");
         const item = await ItemServiceModel.findItem(id);
         if (!item) {
-            throw new ErrData(400, false, "No Data Found!");
-            // return res.status(404).send({ message: "No Data Found!" });
+            throw ErrData(400, "No data found!", false);
         }
-        return res.status(200).send(item);
+        return response(res, 200, true, "Data found!", item);
     }
     catch (error) {
         console.error("Catch error:-", error);
-        res.status(500).send(error);
+        next(error);
     }
 }
 
 // Update items using Object id
-export const updateItemController = async (req: Request, res: Response) => {
+export const updateItemController = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        let reqBody = req.body;
-
-        const updateResult = await ItemServiceModel.updateItem(reqBody);
+        if (req.user.role !== Enum.adminRoleId)
+            throw ErrData(400, "You are not authorized to perform this action!", false);
+        let Body = req.body;
+        const updateResult = await ItemServiceModel.updateItem(Body);
         if (!updateResult) {
-            return res.status(400).send({ message: "Item could not be updated!" });
+            throw ErrData(400, "Item could not be updated!", false);
         }
-        return res.status(200).send({ message: "Item details has been updated successfully!" });
+        return response(res, 200, true, "Item details has been updated successfully!", updateResult);
     }
     catch (error) {
         console.error("Catch error:-", error);
-        res.status(500).send(error);
+        next(error);
     }
 }
 
 // delete items using Object id
-export const deleteItemController = async (req: Request, res: Response) => {
+export const deleteItemController = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        let id: string = (req.query.id + "");
+        if (req.user.role !== Enum.adminRoleId)
+            throw ErrData(400, "You are not authorized to perform this action!", false);
 
+        let id: string = toStrings(req.query.id + "");
         let deleteResult = await ItemServiceModel.deleteItem(id);
         if (!deleteResult) {
-            return res.status(400).send({ message: "Item could not be deleted!" });
+            throw ErrData(400, "Item could not be deleted!", false);
         }
-        return res.status(200).send({ message: "Item has been deleted successfully!" })
+        return response(res, 200, true, "Item has been deleted successfully!", deleteResult);
     }
     catch (error) {
         console.error("Catch error:-", error);
-        res.status(500).send(error);
+        next(error);
     }
 }
